requirements.txt
psycopg2-binary==2.9.9
playwright==1.40.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
